+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import rateMyUnitLogo from './assets/logo-default.svg'
-import './App.css'
+import './styles/App.css'
 
 function App() {
   const [count, setCount] = useState(0)
